logics are seperated now
